csproj-6 Add skeleton class for ManagedObject - add initial implementation of IManagedObject - correct Value property in IManagedObject - it should be a property - renamed maybe owned to maybe managed object
